replace the old event test with the extended test
